testing(bigquery): relax migration testing (#3586)
* testing(bigquery): extend test timeout

BQ migration will occasionally have slow workflows that encroach on the
existing limit.  This PR extends the timeout for the service quickstart
and cleans up output a bit.

If this limit is insufficient, we should likely change the test
constraints to only ensure the workflow is started, and not wait
for a successful completion.  This test isn't valuable enough to
add this much wall time to test runs.

Additional note:  this failure mode will appear (per the service log)
as a form of bucket unavailable.  The cause of this is that we correctly
clean up the GCS bucket as a test resource while the workflow is still
pending, yielding an error that's a red herring.

Fixes: https://github.com/GoogleCloudPlatform/golang-samples/issues/3567

* relax test, reduce to 60s timeout